% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cluster.R
\name{assign_cluster}
\alias{assign_cluster}
\alias{assign_cluster.hierarchical_cluster}
\title{Assign Clusters to Documents/Text Elements}
\usage{
assign_cluster(x, k = approx_k(get_dtm(x)), h = NULL, ...)

\method{assign_cluster}{hierarchical_cluster}(x, k = approx_k(get_dtm(x)),
  h = NULL, ...)
}
\arguments{
\item{x}{a \code{hierarchical_cluster} object.}

\item{k}{The number of clusters (can supply \code{h} instead).  Defaults to
use \code{approx_k} of the \code{\link[tm]{DocumentTermMatrix}} produced
by \code{data_storage}.}

\item{h}{The height at which to cut the dendrograms (determines number of
clusters).  If this argument is supplied \code{k} is ignored.}

\item{\ldots}{ignored.}
}
\value{
Returns an \code{assign_cluster} object; a named vector of cluster
assignments with documents as names.  The object also contains the original
\code{data_storage} object.
}
\description{
Assign clusters to documents/text elements.
}
\examples{
library(dplyr)

x <- with(
    presidential_debates_2012,
    data_store(dialogue, paste(person, time, sep = "_"))
)

hierarchical_cluster(x) \%>\%
    plot(h=.7, lwd=2)

hierarchical_cluster(x) \%>\%
    assign_cluster(h=.7)

hierarchical_cluster(x, method="complete") \%>\%
    plot(k=6)

hierarchical_cluster(x) \%>\%
    assign_cluster(k=6)


x2 <- presidential_debates_2012 \%>\%
    with(data_store(dialogue)) \%>\%
    hierarchical_cluster()

ca <- assign_cluster(x2, k = 55)
summary(ca)

## split text into clusters
get_text(ca)
}

