% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dtm.R
\name{get_dtm}
\alias{get_dtm}
\alias{get_dtm.data_store}
\alias{get_dtm.hierarchical_cluster}
\title{Get a \code{\link[tm]{DocumentTermMatrix}} Stored in a \code{hierarchical_cluster} Object}
\usage{
get_dtm(x, ...)

\method{get_dtm}{data_store}(x, ...)

\method{get_dtm}{hierarchical_cluster}(x, ...)
}
\arguments{
\item{x}{A \code{\link[hclustext]{hierarchical_cluster}} object.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{\link[tm]{DocumentTermMatrix}}.
}
\description{
Extract the \code{\link[tm]{DocumentTermMatrix}} supplied to/produced by a
\code{\link[hclustext]{hierarchical_cluster}} object.
}
\examples{
presidential_debates_2012 \%>\%
    with(data_store(dialogue)) \%>\%
    hierarchical_cluster() \%>\%
    get_dtm()
}

