% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustext-package.R
\docType{package}
\name{hclustext}
\alias{hclustext}
\alias{hclustext-package}
\alias{package-hclustext}
\title{Optimized Hierarchical Clustering for Text Data}
\description{
Optimized tools for clustering text data via hierarchical clustering.
}

