% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cluster.R
\name{summary.assign_cluster}
\alias{summary.assign_cluster}
\title{Summary of an assign_cluster Object}
\usage{
\method{summary}{assign_cluster}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{An assign_cluster object.}

\item{\ldots}{ignored.}
}
\description{
Summary of an assign_cluster object
}

